server/iot_server/src/requirements.txt
awscrt
awsiot
python-dotenv
google-cloud-vision
pillow

